+++ lib/twenv/commands/delete_my_tweets.rb
@@ -21,7 +21,7 @@ class TWEnv::DeleteMyTweets < TWEnv::Command
   def process
     perform_action_on_tweets method(:read_tweets),
                              lambda {|tweet| client.destroy_tweet(tweet)},
-                             lambda {|total| line.print(total == 0 ? "No tweets to delete" : "#{total} tweets deleted") }
+                             lambda {|total| line.empty_line!.print(total == 0 ? "No tweets to delete" : "#{total} tweets deleted") }
   end
 
   private